DFS Working & Test functions in separate class
