+++ app/DFS.h
@@ -24,10 +24,4 @@ using namespace std;
 //QHash<QString, StadiumDistance*> DFS(const StadiumGraph& graph, Stadium* start);
 QVector<StadiumDistance*> DFS(const StadiumGraph& graph, Stadium* start);
 
-// Creates & prints the specified path for DFS
-void runDFS(const StadiumGraph& graph, QString start);
-
-// Runs / calls path for DFS from start stadium
-void testDFS();
-
 #endif // DFS_H
