+++ app/Dijkstra.h
@@ -27,10 +27,4 @@ using namespace std;
 // for multiple searches for the shortest path from the start location.
 QHash<QString, StadiumDistance*> Dijkstra(const StadiumGraph& graph, Stadium* start);
 
-// Creates and prints the specified path for Dijkstra
-void runDijkstra(const StadiumGraph& graph, QString start, QString end);
-
-// Runs / calls various paths for Dijkstra
-void testDijkstra();
-
 #endif // DIJKSTRA_H
